+++ docs/release-notes.rst
@@ -7,9 +7,16 @@ The server is deployed within a Docker image in the Docker registry.
 Version 1.0.2, 6 December 2017
 ---------------------------
 
-* Supports /predict, /transform, /transformDefault, /getBinary, /getBinaryDefault end points
+* Support /predict, /transform, /transformDefault, /getBinary, /getBinaryDefault end points
 
 Version 1.0.3, 3 January 2018
 -------------------------------
 
 * add /putModel end point and add /putProto end point
+
+Version 2.0.0, 22 February 2018
+--------------------------------
+
+* add /operation/{operation} end point 
+
+* support protobuf file with no restriction on the naming of messages and nested input messages.